Finished voting system project
in session of the lab we had to make a voting system, which in it
you can create some votings with questions and some options to choose
from.Every voting had a type which could be single-vote or multi-vote.
In single-vote voter can only vote for only one option, on the other hand
in multi-vote voter can vote for more than 1 option.